Hide Windows error reporting dialog if app crashes even if it was built without Google Breakpad
